api/requirements.txt
alembic==1.4.2
amqp==2.5.2
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
celery==4.4.2
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
email-validator==1.0.5
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
Flask-User==1.0.2.2
Flask-WTF==0.14.3
flower==0.9.4
get==2019.4.13
gunicorn==20.0.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.23.0
more-itertools==8.2.0
numpy==1.18.3
packaging==20.3
pandas==1.0.3
passlib==1.7.2
pluggy==0.13.1
post==2019.4.13
public==2019.4.13
py==1.8.1
pycparser==2.20
pycryptodome==3.9.7
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
query-string==2019.4.13
redis==3.4.1
request==2019.4.13
six==1.14.0
SQLAlchemy==1.3.16
tornado==6.0.4
vine==1.3.0
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.3.1
xlrd==1.2.0
XlsxWriter==1.2.8
zipp==3.1.0
